requirements.txt
discord.py==2.5.2
asyncpraw==7.8.1
python-dotenv==1.0.0
aiohttp==3.9.1
colorama==0.4.6
pytest==7.4.3
pytest-asyncio==0.21.1